+++ packages/markdown/editor/language.ts
@@ -44,7 +44,7 @@ const markdownHighlighting = styleTags({
 });
 const nodeTypes = [
   'strong', 'strongSequence', 'emphasis', 'emphasisSequence', 'strikethrough', 'strikethroughSequence', 
-  'codeText', 'codeFenced', 'table', 'blockQuote',
+  'codeText', 'codeFenced', 'table', 'tableRow', 'blockQuote',
   'heading1', 'heading2', 'heading3', 'heading4', 'heading5', 'heading6',
   'link', 'image', 'resource', 'label', 'labelMarker', 'labelText',
   'inlineFootnote', 'inlineFootnoteMarker', 'inlineFootnoteStartMarker', 'inlineFootnoteEndMarker',