+++ packages/markdown/mdext/rehypePlugins.js
@@ -132,3 +132,17 @@ export function rehypeRawFixPassthroughStitches() {
     }
   })
 }
+
+
+/**
+ * Annotate markdown blocks with line numbers as HTML attributes
+ */
+export function rehypeAnnotateMarkdownPositions() {
+  return tree => visit(tree, 'element', (node, _, parent) => {
+    if (node.position && (parent.type === 'root' || (['li', 'tr', 'figure'].includes(node.tagName)))) {
+      node.properties = node.properties || {};
+      node.properties.dataPosition = JSON.stringify(node.position);
+    }
+  });
+}
+